requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
asn1crypto==0.24.0        # via cryptography
azure-common==1.1.18      # via azure-storage
azure-nspkg==3.0.2        # via azure-storage
azure-storage==0.36.0
certifi==2019.3.9         # via requests
cffi==1.12.2              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.6.1       # via azure-storage
django-queryset-csv==1.0.1
django==1.11.20
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
djangorestframework==3.9.2
faker==1.0.4
idna==2.8                 # via requests
psycopg2==2.7.7
pycparser==2.19           # via cffi
pyjwt==1.7.1              # via djangorestframework-jwt
python-dateutil==2.8.0    # via azure-storage, faker
python-dotenv==0.10.1
pytz==2018.9              # via django
requests==2.21.0
six==1.12.0               # via cryptography, faker, python-dateutil
text-unidecode==1.2       # via faker
unicodecsv==0.14.1        # via django-queryset-csv
urllib3==1.24.1           # via requests
